FCBH-883 Implement Freshchat into Bible.is
It is recommended, if not basically required that the user's restore id for Freshchat be stored on an external system from the device.  If the user were to uninstall/reinstall the app their Freshchat would be broken otherwise as this restore id cannot be reset after initially set per user.